Revert "randomization fix"
This reverts commit 3047e3914b42e2368dd100b7f254cdcbc2ea8aad.
