--- P3R.CostumeFramework/UE4SS/Lua/dingalingus/scripts/battlecheck.txt
@@ -1,1 +1,0 @@
-true
