+++ P3R.CostumeFramework/UE4SS/Lua/mods.txt
@@ -1,5 +1,4 @@
 AlwaysPhysics : 1
-dingalingus : 1
